+++ app/[lang]/business/page.tsx
@@ -9,7 +9,6 @@ import { ContactsSection } from '@/sections/business/ContactsSection';
 
 import { getDictionary, getMetadata } from '@/lib/dictionary';
 
-// import { BASE_URL } from '../layout';
 import { FOR_BUSINESS } from '@/data';
 
 import type { Locale } from '@/i18n.config';
